Updated exif data code
Uploaded all images onto the azure blob server. Then updated the exifDataExtract.py file to now be able to look through all images in the blob and extract the metadata from each image. The code now creates a csv file on the local machine and saves the metadata from each image on the csv.